refactor: Move DTOs from lib/dto to types/
- Moved `lib/dto/spotify.dto.ts` to `types/spotify.ts`
- Moved `lib/dto/workout.dto.ts` to `types/workout.ts`
- Updated all relevant import paths to reflect the new file locations.
- Deleted the now-empty `lib/dto` directory.
- This change aligns with the project's convention of keeping all type definitions in the `types/` directory.
